[brain tests] Burst hashlib from the main file
